backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic
reportlab
python-dotenv
aiofiles